improve performance and coding practices
