+++ packages/suite-base/src/panels/Indicator/constants.ts
@@ -1,0 +1,11 @@
+// SPDX-FileCopyrightText: Copyright (C) 2023-2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)<lichtblick@bmwgroup.com>
+// SPDX-License-Identifier: MPL-2.0
+import { Config } from "./types";
+
+export const DEFAULT_CONFIG: Config = {
+  path: "",
+  style: "bulb",
+  fallbackColor: "#a0a0a0",
+  fallbackLabel: "False",
+  rules: [{ operator: "=", rawValue: "true", color: "#68e24a", label: "True" }],
+};
